# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atelier API
  version: "0.1.0"
  description: |
    REST + SSE API for Atelier (canvas-ai) -- an AI-powered UI design generation platform.
    Supports project management, screen generation (including redesign-from-URL and
    image-to-code), component library CRUD, analytics, and feedback collection.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional Bearer token set via ATELIER_API_KEY environment variable.
        When the server has no key configured, all requests are allowed.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        screens:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              prompt:
                type: string
              deviceType:
                type: string
...
    UIComponent:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        name:
          type: string
        description:
          type: string
        html:
          type: string
        tokens:
          type: object
          properties:
            colors:
              type: array
              items:
                type: string
            fonts:
              type: array
              items:
                type: string
            radius:
              type: boolean
        slots:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        adaptability:
          type: string
        quality:
          type: number
        usageCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List all projects
      security: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectSummary"

    post:
      tags: [Projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 255
                  description: Project title (defaults to "Untitled")
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Payload too large (max 10 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}:
    get:
      tags: [Projects]
      summary: Get project details with screen summaries
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project with screens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Storage not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Screens ─────────────────────────────────────────────────────────────
...
  /api/components:
    get:
      tags: [Components]
      summary: List or search components
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by component category
        - name: q
          in: query
          schema:
            type: string
          description: Free-text search query
      responses:
        "200":
          description: Component list
          content:
            application/json:
              schema:
                type: object
                properties:
                  components:
                    type: array
                    items:
                      $ref: "#/components/schemas/UIComponent"
                  total:
                    type: integer

    post:
      tags: [Components]
      summary: Register a new component
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, name, html]
              properties:
                id:
                  type: string
                category:
                  type: string
                name:
                  type: string
                description:
                  type: string
                html:
                  type: string
                tokens:
                  type: object
                  properties:
                    colors:
                      type: array
                      items:
                        type: string
                    fonts:
                      type: array
                      items:
                        type: string
                    radius:
                      type: boolean
                slots:
                  type: array
                  items:
                    type: string
                variants:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
                source:
                  type: string
                adaptability:
                  type: string
                quality:
                  type: number
      responses:
        "201":
          description: Component created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UIComponent"
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/components/{id}:
    get:
      tags: [Components]
      summary: Get a single component by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Component details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UIComponent"
        "404":
          description: Component not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/components/{id}/customize:
    post:
      tags: [Components]
      summary: Customize a component with an AI instruction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instruction]
              properties:
                instruction:
                  type: string
      responses:
        "200":
          description: Customized HTML
          content:
            application/json:
              schema:
                type: object
                properties:
                  html:
                    type: string
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Component not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/components/generate:
    post:
      tags: [Components]
      summary: AI-generate a new component from a description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, category]
              properties:
                description:
                  type: string
                category:
                  type: string
      responses:
        "201":
          description: Generated component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UIComponent"
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...